backend/requirement.txt
flask
flask-cors
python-dotenv
psycopg2-binary
sqlalchemy
supabase
alembic
flask-migrate
requests
gunicorn
pytest
flask_jwt_extended
passlib
pandas
numpy
openpyxl
werkzeug
cryptography
bcrypt
flask_sqlalchemy